+++ framework/modules/saf_utilities/saf_utility_bessel.c
@@ -25,10 +25,6 @@
  */
  
 #include "saf_utilities.h"
-#include <stdio.h>
-#include <stdlib.h>
-#include <math.h>
-#include <string.h>
 
 /* ========================================================================== */
 /*                            Internal Functions                              */
@@ -71,7 +67,8 @@ static int MSTA1
     N1=N0+5;
     F1=ENVJ(N1,A0)-MP;
     for (IT=1; IT<=20; IT++) {
-        NN=N1-(N1-N0)/(1.0-F0/F1);
+        //NN=N1-(N1-N0)/(1.0-F0/F1);
+		NN = N1-(int)((double)(N1-N0) / (1.0-F0/F1));
         F=ENVJ(NN,A0)-MP;
         if (abs(NN-N1) < 1) goto e20;
         N0=N1;
@@ -114,7 +111,8 @@ static int MSTA2
     N1=N0+5;
     F1=ENVJ(N1,A0)-OBJ;
     for (IT=1; IT<=20; IT++) {
-        NN=N1-(N1-N0)/(1.0-F0/F1);
+        //NN=N1-(N1-N0)/(1.0-F0/F1);
+		NN = N1-(int)((double)(N1-N0)/(1.0-F0/F1));
         F=ENVJ(NN,A0)-OBJ;
         if (abs(NN-N1) < 1) goto e20;
         N0=N1;
